Refactor: FieldError 외부 클래스로 변환 - 처음 목적과 다르게 FieldError inner class를 외부에서 사용하게 됨 - 따라서 외부 클래스로 분리
