+++ src/main/java/com/test/board/api/dto/error/ErrorResponse.java
@@ -1,10 +1,8 @@
 package com.test.board.api.dto.error;
 
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 
-import java.util.ArrayList;
 import java.util.List;
 
 
@@ -47,21 +45,6 @@ public class ErrorResponse {
 
     public static ErrorResponse of(final ErrorCode code) {
         return new ErrorResponse(code);
-    }
-
-    @Getter
-    @AllArgsConstructor
-    public static class FieldError {
-        private String field;
-        private String value;
-        private String reason;
-
-        public static List<FieldError> of(final String field, final String value, final String reason) {
-            List<FieldError> fieldErrors = new ArrayList<>();
-            fieldErrors.add(new FieldError(field, value, reason));
-            return fieldErrors;
-        }
-
     }
 
 }
