+++ src/main/java/com/test/board/api/dto/error/FieldError.java
@@ -1,0 +1,22 @@
+package com.test.board.api.dto.error;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Getter
+@AllArgsConstructor
+public class FieldError {
+    private String field;
+    private String value;
+    private String reason;
+
+    public static List<FieldError> of(final String field, final String value, final String reason) {
+        List<FieldError> fieldErrors = new ArrayList<>();
+        fieldErrors.add(new FieldError(field, value, reason));
+        return fieldErrors;
+    }
+
+}
