+++ src/main/java/com/test/board/api/exception/BusinessException.java
@@ -1,7 +1,7 @@
 package com.test.board.api.exception;
 
 import com.test.board.api.dto.error.ErrorCode;
-import com.test.board.api.dto.error.ErrorResponse;
+import com.test.board.api.dto.error.FieldError;
 import lombok.Getter;
 
 import java.util.ArrayList;
@@ -10,9 +10,9 @@ import java.util.List;
 @Getter
 public class BusinessException extends RuntimeException {
     private final ErrorCode errorCode;
-    private final List<ErrorResponse.FieldError> errors;
+    private final List<FieldError> errors;
 
-    public BusinessException(String message, ErrorCode errorCode, List<ErrorResponse.FieldError> errors) {
+    public BusinessException(String message, ErrorCode errorCode, List<FieldError> errors) {
         super(message);
         this.errorCode = errorCode;
         this.errors = errors;
@@ -24,7 +24,7 @@ public class BusinessException extends RuntimeException {
         this.errors = new ArrayList<>();
     }
 
-    public BusinessException(ErrorCode errorCode, List<ErrorResponse.FieldError> errors) {
+    public BusinessException(ErrorCode errorCode, List<FieldError> errors) {
         super(errorCode.getMessage());
         this.errorCode = errorCode;
         this.errors = errors;
